Switched to floats

+++ EliteAPI/Events/ApproachSettlementInfo.cs
@@ -14,10 +14,10 @@ namespace EliteAPI.Events
         public long MarketId { get; internal set; }
 
         [JsonProperty("Latitude")]
-        public double Latitude { get; internal set; }
+        public float Latitude { get; internal set; }
 
         [JsonProperty("Longitude")]
-        public double Longitude { get; internal set; }
+        public float Longitude { get; internal set; }
 
         internal static ApproachSettlementInfo Process(string json, EliteDangerousAPI api)
         {
