+++ EliteAPI/Events/EngineerCraftInfo.cs
@@ -33,7 +33,7 @@ namespace EliteAPI.Events
         public long Level { get; internal set; }
 
         [JsonProperty("Quality")]
-        public double Quality { get; internal set; }
+        public float Quality { get; internal set; }
 
         [JsonProperty("ExperimentalEffect")]
         public string ExperimentalEffect { get; internal set; }
