+++ EliteAPI/Events/FSDFaction.cs
@@ -21,7 +21,7 @@ namespace EliteAPI.Events
         public string Government { get; internal set; }
 
         [JsonProperty("Influence")]
-        public double Influence { get; internal set; }
+        public float Influence { get; internal set; }
 
         [JsonProperty("Allegiance")]
         public string Allegiance { get; internal set; }
@@ -33,7 +33,7 @@ namespace EliteAPI.Events
         public string HappinessLocalised { get; internal set; }
 
         [JsonProperty("MyReputation")]
-        public double MyReputation { get; internal set; }
+        public float MyReputation { get; internal set; }
 
         [JsonProperty("RecoveringStates", NullValueHandling = NullValueHandling.Ignore)]
         public List<IngState> RecoveringStates { get; internal set; }
