+++ EliteAPI/Events/FuelScoopInfo.cs
@@ -5,10 +5,10 @@ namespace EliteAPI.Events
     public class FuelScoopInfo : EventBase
     {
         [JsonProperty("Scooped")]
-        public double Scooped { get; internal set; }
+        public float Scooped { get; internal set; }
 
         [JsonProperty("Total")]
-        public double Total { get; internal set; }
+        public float Total { get; internal set; }
 
         internal static FuelScoopInfo Process(string json, EliteDangerousAPI api)
         {
