+++ EliteAPI/Events/LoadOutModifier.cs
@@ -8,10 +8,10 @@ namespace EliteAPI.Events
         public string Label { get; internal set; }
 
         [JsonProperty("Value")]
-        public double Value { get; internal set; }
+        public float Value { get; internal set; }
 
         [JsonProperty("OriginalValue")]
-        public double OriginalValue { get; internal set; }
+        public float OriginalValue { get; internal set; }
 
         [JsonProperty("LessIsGood")]
         public long LessIsGood { get; internal set; }
