+++ EliteAPI/Events/ModuleRetrieveInfo.cs
@@ -32,7 +32,7 @@ namespace EliteAPI.Events
         public long Level { get; internal set; }
 
         [JsonProperty("Quality")]
-        public double Quality { get; internal set; }
+        public float Quality { get; internal set; }
 
         [JsonProperty("SwapOutItem")]
         public string SwapOutItem { get; internal set; }
