+++ EliteAPI/Status/StatusWatcher.cs
@@ -13,8 +13,8 @@ namespace EliteAPI.Status
         private EliteDangerousAPI api;
         private FileSystemWatcher statusWatcher;
         private bool InNoFireZone = false;
-        private double JumpRange = -1;
-        private double MaxFuel = -1;
+        private float JumpRange = -1;
+        private float MaxFuel = -1;
         private string GameMode = "";
         private bool InMainMenu = false;
         private string MusicTrack = "";
